refactor: reorganize U020 domain classes into a dedicated package
